Consolidate test vectors, add nextest filter (#288)
