+++ crates/web5/src/credentials/presentation_definition.rs
@@ -238,55 +238,3 @@ impl JsonSchemaBuilder {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashSet;
-
-    use crate::test_helpers::TestVectorFile;
-
-    use super::PresentationDefinition;
-
-    #[derive(Debug, serde::Deserialize)]
-    struct VectorInput {
-        #[serde(rename = "presentationDefinition")]
-        pub presentation_definition: PresentationDefinition,
-        #[serde(rename = "credentialJwts")]
-        pub credential_jwts: Vec<String>,
-    }
-
-    #[derive(Debug, serde::Deserialize)]
-    struct VectorOutput {
-        #[serde(rename = "selectedCredentials")]
-        pub selected_credentials: Vec<String>,
-    }
-
-    #[test]
-    #[ignore] // TODO temporarily ignoring, because web5-spec test vectors use did:key which isn't supported
-    fn test_web5_spec_test_vectors() {
-        let path = "presentation_exchange/select_credentials.json";
-        let vectors: TestVectorFile<VectorInput, VectorOutput> =
-            TestVectorFile::load_from_path(path);
-
-        for vector in vectors.vectors {
-            let presentation_definition = vector.input.presentation_definition;
-            let vc_jwts = vector.input.credential_jwts;
-            let error_msg = format!(
-                "Selected Credential test vector ({}) should not have thrown error",
-                vector.description
-            );
-
-            let selected_credentials = presentation_definition
-                .select_credentials(&vc_jwts)
-                .expect(&error_msg);
-
-            let set1: HashSet<_> = selected_credentials.iter().collect();
-            let set2: HashSet<_> = vector.output.selected_credentials.iter().collect();
-            assert_eq!(
-                set1, set2,
-                "Vectors do not contain the same elements: {}",
-                error_msg
-            );
-        }
-    }
-}
