+++ crates/web5/src/dids/methods/did_jwk.rs
@@ -115,71 +115,3 @@ impl DidJwk {
         }
     }
 }
-
-#[cfg(test)]
-mod web5_test_vectors_did_jwk {
-    use crate::{
-        dids::{
-            data_model::document::Document,
-            resolution::{
-                document_metadata::DocumentMetadata, resolution_metadata::ResolutionMetadata,
-            },
-        },
-        test_helpers::TestVectorFile,
-    };
-
-    #[derive(Debug, PartialEq, serde::Deserialize)]
-    struct VectorOutput {
-        #[serde(rename = "@context")]
-        context: String,
-        #[serde(rename = "didDocument")]
-        did_document: Option<Document>,
-        #[serde(rename = "didDocumentMetadata")]
-        did_document_metadata: DocumentMetadata,
-        #[serde(rename = "didResolutionMetadata")]
-        did_resolution_metadata: ResolutionMetadata,
-    }
-
-    #[test]
-    fn resolve() {
-        let path = "did_jwk/resolve.json";
-        let vectors: TestVectorFile<String, VectorOutput> = TestVectorFile::load_from_path(path);
-
-        for vector in vectors.vectors {
-            let did_uri = vector.input;
-            let resolution_result = super::DidJwk::resolve(&did_uri);
-
-            let all_none = vector.output.did_document_metadata.created.is_none()
-                && vector.output.did_document_metadata.updated.is_none()
-                && vector.output.did_document_metadata.deactivated.is_none()
-                && vector.output.did_document_metadata.next_update.is_none()
-                && vector.output.did_document_metadata.version_id.is_none()
-                && vector
-                    .output
-                    .did_document_metadata
-                    .next_version_id
-                    .is_none()
-                && vector.output.did_document_metadata.equivalent_id.is_none()
-                && vector.output.did_document_metadata.canonical_id.is_none();
-
-            let vector_document_metadata = if all_none {
-                None
-            } else {
-                Some(vector.output.did_document_metadata.clone())
-            };
-
-            assert_eq!(
-                resolution_result.resolution_metadata, vector.output.did_resolution_metadata,
-                "Resolution metadata does not match."
-            );
-            assert_eq!(
-                resolution_result.document, vector.output.did_document,
-                "DID Document does not match."
-            );
-            assert_eq!(
-                resolution_result.document_metadata, vector_document_metadata,
-                "Document metadata does not match."
-            );
-        }
-    }
-}
